test: refactor test code structure and optimize test cases
- Remove redundant code
- Optimize test file organization
