+++ package.json
@@ -48,12 +48,11 @@
     "tinyexec": "^1.1.1"
   },
   "devDependencies": {
-    "@configurajs/vite-plus": "^0.2.2",
+    "@configurajs/vite-plus": "^0.2.4",
     "@types/node": "^22.14.0",
     "@types/semver": "^7.7.1",
     "@typescript/native-preview": "7.0.0-dev.20260410.1",
     "@vitest/coverage-v8": "^4.1.4",
-    "conventional-changelog-angular": "^8.3.1",
     "vite-plus": "catalog:"
   },
   "engines": {
