+++ pnpm-workspace.yaml
@@ -1,10 +1,12 @@
 catalog:
   vite: npm:@voidzero-dev/vite-plus-core@latest
+  vite-plus: ^0.1.18
   vitest: npm:@voidzero-dev/vite-plus-test@latest
-  vite-plus: latest
+
 overrides:
   vite: 'catalog:'
   vitest: 'catalog:'
+
 peerDependencyRules:
   allowAny:
     - vite
